hp calc offset to an int to allow negative offsets

+++ MNoFlinchNoWobble/Config.cs
@@ -12,7 +12,7 @@ namespace MNoFlinchNoWobble
     {
         public bool DisableFlinching;
         public bool DisableWobble;
-        public byte HPCalcOffset;
+        public int HPCalcOffset;
         public List<AntiWobbleExplosionReplacements> AntiWobbleReplacements;
         public void LoadDefaults()
         {
